+++ src/eseq.h
@@ -11,20 +11,15 @@ struct eseq
 };
 
 void                        eseq_destroy(struct eseq const* eseq);
-static inline uint_fast16_t eseq_get(struct eseq const* eseq, uint_fast16_t position,
-                                     uint_fast16_t length);
-static inline uint_fast16_t eseq_length(struct eseq const* eseq);
+static inline uint_fast16_t eseq_get(struct eseq const* eseq, uint_fast32_t position, uint_fast32_t length);
+static inline uint_fast32_t eseq_length(struct eseq const* eseq);
 void                        eseq_setup(struct eseq* eseq, struct imm_seq const* seq);
 
-static inline uint_fast16_t eseq_get(struct eseq const* eseq, uint_fast16_t position,
-                                     uint_fast16_t length)
+static inline uint_fast16_t eseq_get(struct eseq const* eseq, uint_fast32_t position, uint_fast32_t length)
 {
     return matrixu16_get(eseq->code, position, length - seq_code_min_seq(eseq->seq_code));
 }
 
-static inline uint_fast16_t eseq_length(struct eseq const* eseq)
-{
-    return (uint_fast16_t)(matrixu16_nrows(eseq->code) - 1);
-}
+static inline uint_fast32_t eseq_length(struct eseq const* eseq) { return matrixu16_nrows(eseq->code) - 1; }
 
 #endif
